+++ packages/vue-cli-plugin-vue-dash/vueDotLoader.js
@@ -18,6 +18,7 @@ const componentMap = {
 	// Patterns
 	AccessibilityBanner: 'patterns',
 	CookieBanner: 'patterns',
+	CollapsibleList: 'patterns',
 	DataListGroup: 'patterns',
 	DatePicker: 'patterns',
 	FileUpload: 'patterns',
