--- packages/vue-dot/src/patterns/FooterBar/CollapseList/index.ts
@@ -1,3 +1,0 @@
-import CollapseList from './CollapseList.vue';
-
-export default CollapseList;
