+++ packages/vue-dot/src/patterns/FooterBar/CollapsibleList/index.ts
@@ -1,0 +1,3 @@
+import CollapseList from './CollapsibleList.vue';
+
+export default CollapseList;
