+++ packages/vue-dot/src/patterns/FooterBar/CollapsibleList/types.d.ts
